# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRM Backend API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/human": {
            "get": {
                "description": "Получить всех людей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "люди"
                ],
                "summary": "Получить всех людей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Возраст от",
                        "name": "age_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Возраст до",
                        "name": "age_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Пол",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Национальность",
                        "name": "nationaly",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Максимальное количество возвращаемых людей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Число первых пропущенных возвращаемых людей",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_P1xart_effective_mobile_service_internal_controller_v1_response.GetAllHumans"
                        }
                    }
                }
            }
        },
        "/v1/human/": {
            "post": {
                "description": "Создание нового человека",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "люди"
                ],
                "summary": "Создание нового человека",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_P1xart_effective_mobile_service_internal_controller_v1_request.CreateHuman"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/v1/human/{id}": {
            "delete": {
                "description": "Удалить человека по его идентификатору",
                "tags": [
                    "люди"
                ],
                "summary": "Удалить человека по его идентификатору",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор человека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "Обновить человека по его идентификатору. Принимает JSON с обновленными полями",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "люди"
                ],
                "summary": "Обновить человека по его идентификатору",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор человека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_P1xart_effective_mobile_service_internal_controller_v1_request.UpdateHuman"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_P1xart_effective_mobile_service_internal_controller_v1_request.CreateHuman": {
            "type": "object",
            "required": [
                "name",
                "surname"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Vasiliy"
                },
                "potronymic": {
                    "type": "string",
                    "example": "Vasilevich"
                },
                "surname": {
                    "type": "string",
                    "example": "Vasiliev"
                }
            }
        },
        "github_com_P1xart_effective_mobile_service_internal_controller_v1_request.UpdateHuman": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "potronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "github_com_P1xart_effective_mobile_service_internal_controller_v1_response.GetAllHumans": {
            "type": "object",
            "properties": {
                "humans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_P1xart_effective_mobile_service_internal_entity.Human"
                    }
                }
            }
        },
        "github_com_P1xart_effective_mobile_service_internal_entity.Human": {
            "type": "object",
            "properties": {
                "Gender": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "potronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}